22 - The Principle of Least Privilege

+++ app/Http/Controllers/Api/V1/UserController.php
@@ -24,20 +24,19 @@ class UserController extends ApiController
             User::filter($filters)->paginate()
         );
     }
+
     /**
      * Store a newly created resource in storage.
      */
     public function store(StoreUserRequest $request)
     {
-        try {
-            // policy
-            $this->isAble('store', User::class);
-
+        if ($this->isAble('store', User::class)) {
             return new UserResource(User::create($request->mappedAttributes()));
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to create that resource', 401);
         }
+
+        return $this->error('You are not authorized to create that resource', 401);
     }
+
     /**
      * Display the specified resource.
      */
@@ -46,8 +45,10 @@ class UserController extends ApiController
         if ($this->include('tickets')) {
             return new UserResource($user->load('tickets'));
         }
+
         return new UserResource($user);
     }
+
     /**
      * Update the specified resource in storage.
      */
@@ -55,32 +56,37 @@ class UserController extends ApiController
     {
         try {
             $user = User::findOrFail($user_id);
-            // policy
-            $this->isAble('update', $user);
 
-            $user->update($request->mappedAttributes());
+            if ($this->isAble('update', $user)) {
+                $user->update($request->mappedAttributes());
 
-            return new UserResource($user);
+                return new UserResource($user);
+            }
+
+            return $this->error('You are not authorized to update that resource', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update that resource', 401);
         }
     }
+
     public function replace(ReplaceUserRequest $request, $user_id)
     {
         // PUT
         try {
             $user = User::findOrFail($user_id);
-            // policy
-            $this->isAble('replace', $user);
-            $user->update($request->mappedAttributes());
 
-            return new UserResource($user);
+            if ($this->isAble('replace', $user)) {
+                $user->update($request->mappedAttributes());
+
+                return new UserResource($user);
+            }
+
+            return $this->error('You are not authorized to update that resource', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot be found.', 404);
         }
     }
+
     /**
      * Remove the specified resource from storage.
      */
@@ -88,10 +94,14 @@ class UserController extends ApiController
     {
         try {
             $user = User::findOrFail($user_id);
-            // policy
-            $this->isAble('delete', $user);
-            $user->delete();
-            return $this->ok('User successfully deleted');
+
+            if ($this->isAble('delete', $user)) {
+                $user->delete();
+
+                return $this->ok('User successfully deleted');
+            }
+
+            return $this->error('You are not authorized to delete that resource', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot found.', 404);
         }
